+++ Game1/Industries/IndustryConfig.cs
@@ -14,7 +14,7 @@ namespace Game1.Industries
             startingPowerPlantParams = new PowerPlant.GeneralBuildingParams
             (
                 name: "Basic Power Plant",
-                buildingComponentPropors: new List<(Product.Params prodParams, ulong amount)>()
+                buildingComponentPropors: new List<(Product.Params prodParams, uint amount)>()
                 {
                     (prodParams: Product.productParamsDict["Wire"], amount: 4),
                     (prodParams: Product.productParamsDict["Roof Tile"], amount: 1)
@@ -24,7 +24,7 @@ namespace Game1.Industries
             startingStorageParams = new Storage.GeneralBuildingParams
             (
                 name: "Basic Storage",
-                buildingComponentPropors: new List<(Product.Params prodParams, ulong amount)>()
+                buildingComponentPropors: new List<(Product.Params prodParams, uint amount)>()
                 {
                     (prodParams: Product.productParamsDict["Gear"], amount: 4),
                     (prodParams: Product.productParamsDict["Roof Tile"], amount: 1)
@@ -49,7 +49,7 @@ namespace Game1.Industries
                     (
                         name: "Basic Mining",
                         energyPriority: new(value: 20),
-                        buildingComponentPropors: new List<(Product.Params prodParams, ulong amount)>()
+                        buildingComponentPropors: new List<(Product.Params prodParams, uint amount)>()
                         {
                             (prodParams: Product.productParamsDict["Gear"], amount: 4),
                             (prodParams: Product.productParamsDict["Wire"], amount: 1),
